pyproject.toml
[tool.poetry]
name = "cursor-contracts-manager"
version = "0.1.0"
description = "Sistema de gestão de contratos de consultoria"
authors = ["Portal Coddfy Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
# Framework Pyramid
pyramid = "^2.0.2"
pyramid-jinja2 = "^2.10"
waitress = "^3.0.0"

# Banco de dados
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
zope-sqlalchemy = "^3.1"

# API e validação
marshmallow = "^3.20.1"

# Autenticação JWT
pyjwt = "^2.8.0"
bcrypt = "^4.1.1"

# Utilitários
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"
pytz = "^2023.3"
pyramid-tm = "^2.6"
pyramid-openapi3 = "^0.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
ipython = "^8.18.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

